requirements/production.txt
Flask
flask-login
flask-sqlalchemy
flask-httpauth
psycopg2
flask-migrate
click
sqlalchemy_utils
cryptography
gunicorn
